+++ src/hooks.js
@@ -1,4 +1,7 @@
-import { countOccurences, runSSHCommand } from './utils';
+import {
+  countOccurences,
+  runSSHCommand
+} from './utils';
 
 export const hooks = {};
 
@@ -25,15 +28,16 @@ export function registerHook(_hookName, _handler) {
 }
 
 export async function runRemoteHooks(serversConfig, command) {
-  return Promise.all(Object.values(serversConfig).map(server => {
-    return runSSHCommand(server, command)
-      .then(({ output }) => {
+  return Promise.all(Object.values(serversConfig)
+    .map(server => runSSHCommand(server, command)
+      .then(({
+        output
+      }) => {
         console.log(`=> output from ${server.host}`);
         console.log(output);
       })
       .catch(e => {
         console.error(`Error running remote hook command: ${command}`);
         console.error(e);
-      });
-  }));
+      })));
 }