+++ src/plugins/default/index.js
@@ -1,12 +1,14 @@
 import * as _commands from './commands';
 import traverse from 'traverse';
 
-export let commands = _commands;
+export const commands = _commands;
 
 export function scrubConfig(config) {
   if (config.servers) {
+    // eslint-disable-next-line
     config.servers = traverse(config.servers).map(function() {
       if (this.path.length !== 2) {
+        // eslint-disable-next-line
         return;
       }
 
@@ -17,6 +19,8 @@ export function scrubConfig(config) {
           return this.update('password');
         case 'pem':
           return this.update('~/.ssh/pem');
+
+        // no default
       }
     });
   }