+++ src/plugins/docker/assets/init-swarm.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+docker swarm init --advertise-addr <%= host %>