+++ src/plugins/docker/assets/swarm-leave.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+docker swarm leave --force