+++ src/plugins/meteor/index.js
@@ -4,7 +4,7 @@ import traverse from 'traverse';
 
 export const description = 'Deploy and manage meteor apps';
 
-export let commands = _commands;
+export const commands = _commands;
 
 export const validate = {
   meteor: _validator,
@@ -14,6 +14,7 @@ export const validate = {
       // Or the config is telling a different app to handle deployment
       return [];
     }
+
     return _validator(config, utils);
   }
 };
@@ -26,6 +27,7 @@ export function prepareConfig(config) {
   config.app.docker = config.app.docker || {};
   config.app.docker.image = config.app.docker.image || config.app.dockerImage || 'kadirahq/meteord';
   delete config.app.dockerImage;
+
   return config;
 }
 
@@ -34,10 +36,11 @@ function meteorEnabled(api) {
   if (config.app && config.app.type === 'meteor') {
     return true;
   }
+
   return false;
 }
 
-export let hooks = {
+export const hooks = {
   'post.default.setup'(api) {
     if (meteorEnabled(api)) {
       return api.runCommand('meteor.setup');
@@ -73,6 +76,11 @@ export let hooks = {
     if (meteorEnabled(api)) {
       return api.runCommand('meteor.restart');
     }
+  },
+  'post.default.status'(api) {
+    if (meteorEnabled(api)) {
+      return api.runCommand('meteor.status');
+    }
   }
 };
 
@@ -82,8 +90,9 @@ export function scrubConfig(config, utils) {
   }
 
   if (config.app) {
+    // eslint-disable-next-line
     config.app = traverse(config.app).map(function() {
-      let path = this.path.join('.');
+      const path = this.path.join('.');
 
       switch (path) {
         case 'name':
@@ -96,7 +105,7 @@ export function scrubConfig(config, utils) {
 
         case 'env.MONGO_URL':
           if (config.mongo) {
-            let url = this.node.split('/');
+            const url = this.node.split('/');
             url.pop();
             url.push('my-app');
 
@@ -104,9 +113,25 @@ export function scrubConfig(config, utils) {
           }
 
           return this.update(utils.scrubUrl(this.node));
+
+        // no default
       }
     });
   }
 
   return config;
+}
+
+export function swarmOptions(config) {
+  if (config && config.app && config.app.type === 'meteor') {
+    return {
+      labels: Object.keys(config.app.servers).reduce((result, server) => {
+        result[server] = {
+          [`mup-app-${config.app.name}`]: 'true'
+        };
+
+        return result;
+      }, {})
+    };
+  }
 }