+++ src/plugins/proxy/assets/proxy-setup.sh
@@ -7,8 +7,8 @@ sudo mkdir -p /opt/<%= name %>/config
 sudo mkdir -p /opt/<%= name %>/config/vhost.d
 sudo mkdir -p /opt/<%= name %>/config/html
 
-touch /opt/<%= name %>/config/shared-config.sh
-touch /opt/<%= name %>/config/env.list
-touch /opt/<%= name %>/config/env_letsencrypt.list
+sudo touch /opt/<%= name %>/config/shared-config.sh
+sudo touch /opt/<%= name %>/config/env.list
+sudo touch /opt/<%= name %>/config/env_letsencrypt.list
 
 sudo chown ${USER} /opt/<%= name %> -R