+++ src/plugins/proxy/command-handlers.js
@@ -1,5 +1,7 @@
+import chalk from 'chalk';
 import { clone } from 'lodash';
 import debug from 'debug';
+import fs from 'fs';
 import nodemiral from 'nodemiral';
 
 const log = debug('mup:module:proxy');
@@ -16,6 +18,7 @@ export function logs(api) {
 
   const args = api.getArgs().slice(1);
   const sessions = api.getSessions(['app']);
+
   return api.getDockerLogs(PROXY_CONTAINER_NAME, sessions, args);
 }
 
@@ -50,6 +53,7 @@ export function setup(api) {
   }
 
   const list = nodemiral.taskList('Setup proxy');
+  const domains = config.domains.split(',');
 
   list.executeScript('Setup Environment', {
     script: api.resolvePath(__dirname, 'assets/proxy-setup.sh'),
@@ -60,13 +64,39 @@ export function setup(api) {
 
   list.copy('Pushing the Startup Script', {
     src: api.resolvePath(__dirname, 'assets/templates/start.sh'),
-    dest: '/opt/' + PROXY_CONTAINER_NAME + '/config/start.sh',
+    dest: `/opt/${PROXY_CONTAINER_NAME}/config/start.sh`,
     vars: {
       appName: PROXY_CONTAINER_NAME,
       letsEncryptEmail: config.ssl ? config.ssl.letsEncryptEmail : null
     }
   });
 
+  let nginxServerConfig = '';
+  if (config.nginxServerConfig) {
+    nginxServerConfig = fs.readFileSync(
+      api.resolvePath(api.getBasePath(), config.nginxServerConfig)
+    ).toString('utf8');
+  }
+
+  let nginxLocationConfig = '';
+  if (config.nginxLocationConfig) {
+    nginxLocationConfig = fs.readFileSync(
+      api.resolvePath(api.getBasePath(), config.nginxLocationConfig)
+    ).toString('utf8');
+  }
+
+  list.executeScript('Pushing Nginx Config', {
+    script: api.resolvePath(__dirname, 'assets/nginx-config.sh'),
+    vars: {
+      hasServerConfig: config.nginxServerConfig,
+      hasLocationConfig: config.nginxLocationConfig,
+      serverConfig: nginxServerConfig,
+      locationConfig: nginxLocationConfig,
+      domains,
+      proxyName: PROXY_CONTAINER_NAME,
+      clientUploadLimit: config.clientUploadLimit || '10M'
+    }
+  });
 
   list.executeScript('Cleaning Up SSL Certificates', {
     script: api.resolvePath(__dirname, 'assets/ssl-cleanup.sh'),
@@ -95,7 +125,7 @@ export function setup(api) {
       vars: {
         appName,
         proxyName: PROXY_CONTAINER_NAME,
-        domains: config.domains.split(',')
+        domains
       }
     });
   }
@@ -127,7 +157,7 @@ export function reconfigShared(api) {
 
   list.copy('Sending shared variables', {
     src: api.resolvePath(__dirname, 'assets/templates/shared-config.sh'),
-    dest: '/opt/' + PROXY_CONTAINER_NAME + '/config/shared-config.sh',
+    dest: `/opt/${PROXY_CONTAINER_NAME}/config/shared-config.sh`,
     vars: {
       httpPort: shared.httpPort,
       httpsPort: shared.httpsPort,
@@ -139,7 +169,7 @@ export function reconfigShared(api) {
 
   list.copy('Sending proxy environment variables', {
     src: api.resolvePath(__dirname, 'assets/templates/env.list'),
-    dest: '/opt/' + PROXY_CONTAINER_NAME + '/config/env.list',
+    dest: `/opt/${PROXY_CONTAINER_NAME}/config/env.list`,
     vars: {
       env: env || {}
     }
@@ -149,13 +179,14 @@ export function reconfigShared(api) {
 
   list.copy('Sending let\'s encrypt environment variables', {
     src: api.resolvePath(__dirname, 'assets/templates/env.list'),
-    dest: '/opt/' + PROXY_CONTAINER_NAME + '/config/env_letsencrypt.list',
+    dest: `/opt/${PROXY_CONTAINER_NAME}/config/env_letsencrypt.list`,
     vars: {
       env: envLetsEncrypt || {}
     }
   });
 
   const sessions = api.getSessions(['app']);
+
   return api.runTaskList(list, sessions, {
     series: true,
     verbose: api.verbose
@@ -180,6 +211,7 @@ export function start(api) {
   });
 
   const sessions = api.getSessions(['app']);
+
   return api.runTaskList(list, sessions, {
     series: true,
     verbose: api.getVerbose()
@@ -199,5 +231,95 @@ export function stop(api) {
   });
 
   const sessions = api.getSessions(['app']);
+
   return api.runTaskList(list, sessions, { verbose: api.getVerbose() });
+}
+
+export async function nginxConfig(api) {
+  log('exec => mup proxy nginx-config');
+
+  const command = `docker exec ${PROXY_CONTAINER_NAME} cat /etc/nginx/conf.d/default.conf`;
+  const { servers, app } = api.getConfig();
+  const serverObjects = Object.keys(app.servers)
+    .map(serverName => servers[serverName]);
+
+
+  await Promise.all(
+    serverObjects.map(server =>
+      api.runSSHCommand(server, command)
+    )
+  ).then(results => {
+    results.forEach(({ host, output }) => {
+      console.log(`===== ${host} ======`);
+      console.log(output);
+    });
+  });
+}
+
+export async function status(api) {
+  const config = api.getConfig();
+  const servers = Object.keys(config.app.servers)
+    .map(key => config.servers[key]);
+  const lines = [];
+  const overallColor = 'green';
+
+  const collectorConfig = {
+    nginxDocker: {
+      command: `docker inspect ${PROXY_CONTAINER_NAME} --format "{{json .}}"`,
+      parser: 'json'
+    },
+    letsEncryptDocker: {
+      command: `docker inspect ${PROXY_CONTAINER_NAME}-letsencrypt --format "{{json .}}"`,
+      parser: 'json'
+    },
+    certificateExpire: {
+      command: `cd /opt/${PROXY_CONTAINER_NAME}/mounted-certs && find . -name '*.chain.pem' -exec echo '{}' \\; -exec openssl x509 -enddate -noout -in '{}' \\;`,
+      parser(stdout, code) {
+        if (code === 0) {
+          return stdout.split('\n').reduce((result, item, index, items) => {
+            if (!(index % 2) && item.trim() !== '') {
+              result[item.slice(2)] = items[index + 1].split('=')[1];
+            }
+
+            return result;
+          }, {});
+        }
+
+        return null;
+      }
+    }
+  };
+
+  const serverInfo = await api.getServerInfo(servers, collectorConfig);
+
+  Object.values(serverInfo).forEach(
+    ({ _host, nginxDocker, letsEncryptDocker, certificateExpire }) => {
+      lines.push(` - ${_host}:`);
+      lines.push('   - NGINX:');
+      lines.push(`     - Status: ${nginxDocker ? nginxDocker.State.Status : 'Stopped'}`);
+
+      if (nginxDocker) {
+        lines.push('     - Ports:');
+        Object.keys(nginxDocker.NetworkSettings.Ports || {}).forEach(key => {
+          if (key === '443/tcp') {
+            lines.push(`       - HTTPS: ${nginxDocker.NetworkSettings.Ports[key][0].HostPort}`);
+          } else if (key === '80/tcp') {
+            lines.push(`       - HTTP: ${nginxDocker.NetworkSettings.Ports[key][0].HostPort}`);
+          }
+        });
+      }
+
+      lines.push('   - Let\'s Encrypt');
+      lines.push(`     - Status: ${letsEncryptDocker ? letsEncryptDocker.State.Status : 'Stopped'}`);
+
+      if (certificateExpire && certificateExpire.length > 0) {
+        lines.push('     - Certificates');
+        Object.keys(certificateExpire).forEach(key => {
+          lines.push(`       - ${key}: ${certificateExpire[key]}`);
+        });
+      }
+    });
+
+  console.log(chalk[overallColor]('\n=> Reverse Proxy Status'));
+  console.log(lines.join('\n'));
 }