+++ src/updates.js
@@ -1,14 +1,15 @@
-import Npm from 'silent-npm-registry-client';
 import boxen from 'boxen';
 import chalk from 'chalk';
 import debug from 'debug';
+import Npm from 'silent-npm-registry-client';
 import pkg from '../package.json';
 
-let log = debug('mup:updates');
+const log = debug('mup:updates');
 
 export default function() {
   log('checking for updates');
-  return new Promise(function(resolve) {
+
+  return new Promise(resolve => {
     const params = {
       timeout: 1000,
       package: pkg.name,
@@ -17,9 +18,10 @@ export default function() {
 
     const npm = new Npm();
     const uri = 'https://registry.npmjs.org/npm';
-    npm.distTags.fetch(uri, params, function(err, res) {
+    npm.distTags.fetch(uri, params, (err, res) => {
       if (err) {
         resolve();
+
         return;
       }
 
@@ -28,7 +30,8 @@ export default function() {
       const npmVersion = res.latest;
       const nextVersion = res.next;
 
-      const local = pkg.version.split('.').slice(0, 3).map(n => Number(n.split('-')[0]));
+      const local = pkg.version.split('.').slice(0, 3)
+        .map(n => Number(n.split('-')[0]));
       const remote = npmVersion.split('.').map(n => Number(n.split('-')[0]));
       const next = nextVersion.split('.').map(n => Number(n.split('-')[0]));
       next.push(nextVersion.split('.')[2].split('beta')[1]);
@@ -56,8 +59,8 @@ export default function() {
       }
 
       if (available) {
-        let version = showStable ? npmVersion : nextVersion;
-        let command = showStable ? 'npm i -g mup' : 'npm i -g mup@next';
+        const version = showStable ? npmVersion : nextVersion;
+        const command = showStable ? 'npm i -g mup' : 'npm i -g mup@next';
 
         let text = `update available ${pkg.version} => ${version}`;
         text += `\nTo update, run ${chalk.green(command)}`;